working on userAccount mvc login -> userDTO rest login

+++ Bowwow/src/main/java/ga/bowwow/service/user/VO/UserAccount.java
@@ -17,7 +17,7 @@ public class UserAccount {
 	String phone;
 	String password;
 	String status;
-	int point;
+	long point;
 	String coupons;
 
 	@Override
